pyproject.toml
[project]
name = "openreview-dl"
version = "0.1.0"
description = "Download OpenReview paper reviews and rebuttals as formatted documents"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "BSD-3-Clause"}
authors = [
    {name = "Dariush Wahdany"}
]
keywords = ["openreview", "reviews", "academic", "papers", "downloader"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "cryptography>=44.0.2",
    "markdown>=3.7",
    "openreview-py>=1.47.1",
    "pypandoc>=1.15",
    "urllib3>=2.0.0",
]

[project.scripts]
openreview-dl = "openreview_dl.cli:main"

[project.urls]
Homepage = "https://github.com/dwahdany/openreview-dl"
Repository = "https://github.com/dwahdany/openreview-dl"
Issues = "https://github.com/dwahdany/openreview-dl/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/openreview_dl"]
